Refactored "tbl_structure.php" by Creating "display_structure.lib.php" file to separate presentation layer from the logic layer.
